Use strings as alert Hash keys to prevent problems with JSON (de)serialisation
When calling render the flash arrives to the next action as it is, without
any further processing. However, when invoking redirect_to the flash is
saved into the session. As we are using Rails' default session storage,
CookieStore, this means that on redirect the session will be saved into a
cookie and retrieved again later. When storing data into cookies the
default serialisation mechanism is :json. JSON serialization represents
hashes as Stringa before encryption. After it is decrypted, keys are
deserialized as a string.

This commit accounts for this behaviour by replacting the symbol keys in
the Hashes used for alerts with string keys.

For more information see: https://github.com/rails/rails/issues/33447