Bump sassc and fix LoadErrors
There seems to be a portability issue in the sassc gem that attempts to execute illegal instruction which subsequently manifest themselves as Ruby LoadErrors. The workaround for this problem is to uninstall the gem (if already installed) and then reinstall it without tuning the native extension code for the CPU architecture. This leads to marginally less efficient code, however this is acceptable since libsass is only used when precompiling assets.

More information is available at the following GitHub issue page for the sassc gem:
https://github.com/sass/sassc-ruby/issues/146