Refactor deprecations in booking validator
As of approximately 13:15 on October 15, Camdram has deprecated the top
level venue field from the JSON returned by the show endpoint. In this
commit we thus appropriately refactor the validator which deals with
checking that a show is authorised to make bookings in the room
currently associated with the model so that it iterates over each
performance and checks the venue present in the JSON there.