Switch to Azure Blob Storage for caching build artefacts
This commit switches the backend storage used for caching build
artefacts from Backblaze B2 to Azure Blob Storage. This should lead to
large improvements in project build time (I have observed a decrease of
about 50%) since the Storage Account in use is hosted in the same Azure
Region as that in which the GitHub Actions runner executes the CI suite.

Signed-off-by: Charlie Jonas <charlie@charliejonas.co.uk>