Fix bug in sitemap generator
The `find_each` method that was used previously had the side effect fo
reversing the sort order of the SQL query. As the memory impact of
loading 250 bookings should not be severe, this commit removes the
`find_each` method to ensure the shows are fed to the generator in the
correct order.

Signed-off-by: Charlie Jonas <charlie@charliejonas.co.uk>