Add user-show and user-society relationships
This commit is in preparation for a rewrite of the Room Booking user
permission system. The intention is to use the many-to-many
relationship between users and the shows & societies on Camdram that
they administer that is introduced by this commit to determine
whether or not a user is authorised to make bookings.

This will eventually allow for much quicker page load times as requests
to Camdram do not have to be made to determine the list of a user's
authorised shows/societies, at the expense of potentially stale
out-of-date data.

Signed-off-by: Charlie Jonas <charlie@charliejonas.co.uk>