Memoize CamdramShow names as part of the database record
Shows are frequently cancelled or otherwise deleted from Camdram, which
leaves an orphaned record on Roombooking which, in order to preserve
historic booking data, we may not necessarily want to delete.

To combat this, we can indefinitely memoize the name of the show as part
of the record that's saved to the database and then update this in a
background job.

Signed-off-by: Charlie Jonas <charlie@charliejonas.co.uk>